Combine registries+filters to avoid cyclic imports via Handler
