+++ kopf/structs/registries.py
@@ -16,7 +16,6 @@ import functools
 from types import FunctionType, MethodType
 from typing import MutableMapping, NamedTuple, Text, Optional, Tuple, Callable, Mapping
 
-from kopf.structs import filters
 from kopf.structs import resources as resources_
 
 
@@ -136,12 +135,12 @@ class SimpleRegistry(BaseRegistry):
             if handler.event is None or handler.event == cause.event:
                 if handler.initial and not cause.initial:
                     pass  # ignore initial handlers in non-initial causes.
-                elif filters.match(handler=handler, body=cause.body, changed_fields=changed_fields):
+                elif match(handler=handler, body=cause.body, changed_fields=changed_fields):
                     yield handler
 
     def iter_event_handlers(self, resource, event):
         for handler in self._handlers:
-            if filters.match(handler=handler, body=event['object']):
+            if match(handler=handler, body=event['object']):
                 yield handler
 
     def iter_extra_fields(self, resource):
@@ -152,7 +151,7 @@ class SimpleRegistry(BaseRegistry):
     def requires_finalizer(self, resource, body):
         # check whether the body matches a deletion handler
         for handler in self._handlers_requiring_finalizer:
-            if filters.match(handler=handler, body=body):
+            if match(handler=handler, body=body):
                 return True
 
         return False
@@ -274,3 +273,38 @@ def set_default_registry(registry: GlobalRegistry):
     """
     global _default_registry
     _default_registry = registry
+
+
+def match(handler, body, changed_fields=None):
+    return (
+        (not handler.field or _matches_field(handler, changed_fields or [])) and
+        (not handler.labels or _matches_labels(handler, body)) and
+        (not handler.annotations or _matches_annotations(handler, body))
+    )
+
+
+def _matches_field(handler, changed_fields):
+    return any(field[:len(handler.field)] == handler.field for field in changed_fields)
+
+
+def _matches_labels(handler, body):
+    return _matches_metadata(handler=handler, body=body, metadata_type='labels')
+
+
+def _matches_annotations(handler, body):
+    return _matches_metadata(handler=handler, body=body, metadata_type='annotations')
+
+
+def _matches_metadata(handler, body, metadata_type):
+    metadata = getattr(handler, metadata_type)
+    object_metadata = body.get('metadata', {}).get(metadata_type, {})
+
+    for key, value in metadata.items():
+        if key not in object_metadata:
+            return False
+        elif value is not None and value != object_metadata[key]:
+            return False
+        else:
+            continue
+
+    return True
